ztp: update default subscription CRs
Update each subscription CR with default "installPlanApproval"
set to Manual. With this configuration, installPlan will not be
installed automatically. TALO needs to find the right installPlan
and approve it.

Also add check of "AtLatestKnown" status in subscriptions, so
a new version of z-stream operator pushed to the registry will
cause the sub policy to be non-compliant then TALO is able to
know when to start upgrade and trigger upgrade.

Signed-off-by: Angie Wang <angwang@redhat.com>