cnf-tests: add numaresources e2e suite
the NUMA resources suite implements e2e testing of the NUMA
aware scheduler.
Testing the NUMA aware scheduler is a complex business
which requires careful setup, external components (device plugins)
only for test and quite a lot of invasive actions on the cluster.

In order to minimize the disruption, we add the NUMA testsuite
in a separate executable suite (e.g. NOT in cnftests)
and we add extra code in the entry point to skip the suite
alltogether unless the focus contains the numa resources identifier.

Signed-off-by: Francesco Romani <fromani@redhat.com>